cmd/cgo, runtime: add checks for passing pointers from Go to C
This implements part of the proposal in issue 12416 by adding dynamic
checks for passing pointers from Go to C.  This code is intended to be
on at all times.  It does not try to catch every case.  It does not
implement checks on calling Go functions from C.

The new cgo checks may be disabled using GODEBUG=cgocheck=0.

DO NOT SUBMIT
until proposal 12416 is officially adopted.

Update #12416.

Change-Id: I48de130e7e2e83fb99a1e176b2c856be38a4d3c8