Update patch set 2
Patch Set 2:

> These are English specific.  This is not the way.  This is
 > something that should eventually be covered by https://golang.org/issue/12750
 > (I hope).

Yeah I read the design doc, but it's a long way out, doesn't seem to cover support in the core time module, and parsing is explicitly out-of-scope.

In a post-#12750 world would time/format.go disappear entirely?

I thought that this was a pretty minor change which solves a problem now and doesn't really make anything worse

Patch-set: 2