Update patch set 3
Patch Set 3:

R=austin

I'll let Austin decide if this would be too intrusive right now. He seems to still be working on stack barriers on both tip and the Go 1.5 release branch. It might be easier to not rename these for a bit.

(Austin, FWIW, the other numbered suffix files in src/runttime have been merged or renamed.  stack is the notable one remaining)

Patch-set: 3