Update patch set 3
Patch Set 3:

If you don't mind, I'd like to wait until https://go-review.googlesource.com/#/c/15995 is in, since that has to get cherry-picked to the 1.5 branch and that will be much easier if these files haven't been shuffled around. I believe https://go-review.googlesource.com/#/c/15996 is the only other CL I have that this would interact with, and I'm happy to rebase that.

Patch-set: 3