Update patch set 1
Patch Set 1:

(3 comments)

I really like the idea of this and the space wins are good and getting rid of _MSpanListHead is great, but the unsafe embedding of the mspanlink is gnarly.

mSpanList is actually pretty well abstracted, so this could be completely type-safe. For example, mSpanList could be

type mSpanList struct {
    head, tail *mspan
}

and the fields in mspan could be next, prev *mspan just like they are now. This would make the mSpanList a regular doubly-linked list instead of a ring. This slightly complicates a few operations, but eliminates the need for a complicated "next" dance (which may have performance implications) and keeps everything entirely type-safe. I think the only tricky bit would be mSpanList_Remove, which would now require the list. I think that's available everywhere it's used, but it's not entirely obvious.

Patch-set: 1
Label: Code-Review=0