Update patch set 1
Patch Set 1:

(1 comment)

> I really like the idea of this and the space wins are good and
 > getting rid of _MSpanListHead is great, but the unsafe embedding of
 > the mspanlink is gnarly.
 > 
 > mSpanList is actually pretty well abstracted, so this could be
 > completely type-safe. For example, mSpanList could be
 > 
 > type mSpanList struct {
 > head, tail *mspan
 > }
 > 
 > and the fields in mspan could be next, prev *mspan just like they
 > are now. This would make the mSpanList a regular doubly-linked list
 > instead of a ring. This slightly complicates a few operations, but
 > eliminates the need for a complicated "next" dance (which may have
 > performance implications) and keeps everything entirely type-safe.
 > I think the only tricky bit would be mSpanList_Remove, which would
 > now require the list. I think that's available everywhere it's
 > used, but it's not entirely obvious.

I'd love to make it completely type-safe, and I made several attempts at that, but mSpanList_Remove was always the hiccup as you point out.  I'm at least still not familiar enough with the runtime's memory management code to be able to confidently identify what list a span will be a member of for all mSpanList_Remove calls.  Looking again, the ones I feel particularly unsure about are
mHeap_FreeSpanLocked, mCentral_UncacheSpan, and mCentral_FreeSpan.

Brainstorming: Do you think it's worthwhile to add a *mspanlist field to mspan?  If not long-term, it could at least provide a short-term sanity check that we're passing the correct list to mSpanList_Remove.

Patch-set: 1