Update patch set 1
Patch Set 1:

> I'll say LGTM, but that is mysterious code.  Do you think the
 > intent is to ensure that there is a "there" in case someone
 > branches to some special case of a slice?

I'm doing some archaeology on that statement and it seems unneeded at best.  It used to be in a lot of other cases in gen_as_init and it has disappeared from all except this case (the slice case).  I'll poke around some more and see if I can figure out why it was originally added.

Patch-set: 1