Update patch set 2
Patch Set 2:

> (1 comment)


you are right. The test depends on only compiled .a files.
And, this CL is wrong - the bug is the .a files were not in the pkg/android_arm directories any more (maybe due to buildmode pie). Will revert this and send the right fix.

Patch-set: 2
Reviewer: Gerrit User 5190 <5190@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0