Update patch set 1
Patch Set 1:

Thanks, but no thanks.

This CL seems to be eliminating one instance of 'i := (j-1)/2' and perhaps the occasional 'j = j2'. The dominant cost of this algorithm is the calls to Less, not these word-level assignments, and the new and old code make the same number of those calls. I expect the benefit of these optimizations is not measurable, and therefore they are not worth the churn and opportunity for new bugs.

Patch-set: 1