Update patch set 2
Patch Set 2:

(1 comment)

I appreciate the difficulty with zeroed structs and json/xml, especially around not being able to mark a time.Time omitempty. 

The problem is that there is also difficulty in defining which structs should be omitted and which should not. The mistake in the description identified below (zero vs empty) is exactly the point. 

Is it really just zeroed structs that should be omitted? Or all structs containing nothing but empty fields? Or only those structs that - due to unexported fields or themselves having omitempty or ignore on fields of their own - would JSON-format as {}? Or something even more subtle? I don't know the answer here.

I kind of like the elegance and directness of "omitempty for structs means don't emit it if it would encode as {}". The implementation is easy: encode it and then backspace over it if it comes out empty. But that doesn't solve the problem for types implementing json.Marshaler, like time.Time.

Patch-set: 2