Update patch set 1
Patch Set 1:

I think the code diff may be correct but the pseudo-code in the commit message mystifies me.

What's true is that after

	x1, isNil1 := indirect(x)
	x2, isNil2 := indirect(x1)

x1==x2 and isNil1==isNil2. That is, once you've called indirect once, the result doesn't change if you call it again. But it's not true that isNil2==false is guaranteed. And I don't understand the [0] at all.

Also the extra call is redundant, but calling it excessive is an exaggeration.

I think this CL is fine, but please just simplify the commit message to just one line:

	text/template: remove redundant indirect

Thanks.

Patch-set: 1
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0