Update patch set 8
Patch Set 8:

I haven't studied the code, but my understanding is as follows.

Today, encoding/json buffers the entire output of Encode in memory and delivers it to the underlying io.Writer in a single Write.

This CL introduces a bufio.Writer, so that the output can be delivered to the underlying io.Writer in multiple smaller Writes, reducing the need for quite as much intermediate memory.

As noted in the comment Brad linked, the current "buffer until the very end" means that if you call json.Encode on something that is not okay to encode, even if that fact is found out only very late in the encoding process, no output is written to the underlying writer. If you introduce buffering and therefore earlier, smaller writes, you give up this guarantee. 

That is a semantic change and requires some kind of opt-in mechanism.

Thanks.

Patch-set: 8
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0