Update patch set 2
Patch Set 2:

(1 comment)

New patchset fixes up the terminology.

It's true that this doesn't cover the case of unexported fields.  I initially felt that the tradeoff of the simplicity of the implementation outweighed the fact that it didn't cover all cases, but maybe that's not the right thing to do.

Perhaps a better way is to recursively call isEmptyValue() on each exported, non-ignored field in the struct.  That wouldn't take omitempty into account for structs deeper in the hierarchy though.

I like the idea of omitting things that encode as "{}" but not covering time.Time is an issue.  MarshalJSON could accept an io.EOF-like error value indicating if a value is omittable and time.Time's implementation would return it if IsZero() returned true, but that feels gross to me.

Patch-set: 2