net/http, net/http/httptrace: sketch of Transport tracing
Sketch CL with notes from a meeting discussing tracing http client
requests and getting more timing and details out of it.

This adds a new package net/http/httptrace where we can hide this
rarely-used stuff, so as to not majorly pollute the net/http package.
This only adds a new optional pointer field Trace
*httptrace.ClientTrace to the Transport struct.

DO NOT SUBMIT

Updates #12580

Change-Id: I173b49c73be0df6ffb524b43812401be544a3aec