Update patch set 3
Patch Set 3:

so it looks like this breaks TestLinuxSendfile in net/http/fs_test.go because the request to http://localhost/quit is immediately terminated by the server without sending any response (the child helper process does an os.Exit(0) when it receives such a request).

This causes the client to retry (which I think is the intended behavior of this CL), but there is no longer anything accepting new conns from the listener. Therefore the test hangs until it times out.

How do you think this should be fixed?

Patch-set: 3
Label: Code-Review=0