Update patch set 14
Patch Set 14:

BTW I like keeping it in runtime/internal/atomic not internal/atomic for now. We can always broaden later, but for now this is about the runtime, and I don't want to make it any easier for the rest of the standard library to reach into the runtime's guts.

Patch-set: 14
Label: Code-Review=0