Update patch set 7
Patch Set 7:

> Looks like a flaky test.
 > 
 > Can you look into it and deflake? Be sure to test with go test
 > -run=Foo -count=1000 and with and without -race. Maybe also vary
 > -cpu between 1 and 4 (it defaults to the number of CPUs available).

I've spent a couple hours trying a few different approaches here. Unfortunately, the best I've been able to get is something that passes 100% of the time without -race and probably 90% of the time with -race. When it fails, it's because I see an errClosed from the Client:

	transport_test.go:2458: Get http://127.0.0.1:61468/block: net/http: transport closed before response was received


I'm out of ideas at the moment for how to take this test across the finish line.

Patch-set: 7
Label: Code-Review=0