compress/flate: improve flate speed by simplifying copyHist
The copyHist function calls the forwardCopy function, which incurs
the cost of a function call. Also, the forwardCopy function is
written as a generic forwardCopy since it can't reason about the
inputs. Since copyHist has information about copy sections, we can
fold the forwardCopy logic into copyHist in a way that is both
faster and also simpler to understand (we remove 2 levels of loops).

Change-Id: If7da0adf2b0ecab3794cbaa9c1a6d2febe4de456