Update patch set 3
Patch Set 3:

Separating it out hurts the performance since the function cannot be inlined.

I originally came up with this approach when I was working on a Brotli decoder. In my approach, everything related to LZ77 dictionaries was placed into a single struct to deal with that logic. So, instead of testing a single copy function, I heavily tested the LZ77 logic (which includes that forward copy).

PTAL (test has 100% coverage on dict_decoder.go):
* https://github.com/dsnet/compress/blob/brotli/brotli/dict_decoder.go
* https://github.com/dsnet/compress/blob/brotli/brotli/dict_decoder_test.go

With copyHist being a method on decompressor, it is not easy to just test copyHist. If we want, we can just clean up the code and use something similar (if not identical) to my dictDecoder struct. Note that everything in dict_decoder.go is useful for compress/flate except for the LastBytes method.

Patch-set: 3