Update patch set 3
Patch Set 3:

inflate is definitely undertested, considering how important it is, though this CL doesn't really change that.

Regarding this CL, it relies on values in the decompressor to have "sane" values.

wrPos: relies on f.hp >= 0 and f.hp < len(f.hist), otherwise crash.
wrEnd: if weEnd somehow ends up being negative we crash.
rdPos: if f.copyDist is negative it could become >= len(hist), crash.
f.copyLen: We should probably check what state we leave it in, unless we can prove it is valid (and what is that?)

I don't know inflate in the same detail as deflate, so I don't know if the assertions above are 100% safe. 
Unless it can be proven that none of these conditions can occur, there should be a check for them.

Out of curiosity. Have you tried optimizing the case where "f.CopyDist == 1". 
You will encounter this a lot when dealing with long repetitions of the same value. 
If f.copyLen is big, that will be very suboptimal, since it will do a lot of single byte copies, when the operation is basically a memset with the same value.

Great work!

Patch-set: 3
Reviewer: Gerrit User 6545 <6545@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0