Update patch set 1
Patch Set 1:

Trybots are happy, I'm happy.
For the record, the reason the documentation is the way it is is to avoid having to expand the API surface and confusing programmers with the concept of pointers that point to the start of an object and interior pointers.

Patch-set: 1
Label: Code-Review=0