Update patch set 17
Patch Set 17:

(2 comments)

amd64 works and is in a pretty good state!

The latest patch fixes the nuttiness in cmd/go/pkg.go and fixes the issues with the race detector test.

Next up: getting 386 and arm64 to work by replacing atomic stubs with functions that jump to the stubbed function.

Patch-set: 17
Label: Code-Review=0