runtime: inform windows of encoding before writing to console.
The Windows console is usually set to a non-unicode encoding.
Inform windows of the proper encoding about to be written.

This adds at one new syscall for every print to without a
console and four new syscalls for every print to with an
attached console. The Windows console is extreemly slow so
the extra syscalls won't be noticable. It may be possible to
cache the detection of the console but I'm unsure if that can
be done safely.

Fixes #7864

Change-Id: Id13369352aeccac8387876f0b911e383c543c28e