Update patch set 1
Patch Set 1:

> If there a test case or an issue reference for this change ?

There is not, it's just something I noticed.  I don't know if there is a way to write a test for this, since the Darwin ARM environments are so limited.  To be honest there may not even be a way to detect it.  Still, I think it's clearly the right thing to do.

Patch-set: 1