cmd/vet: check for duplicate json struct field tags
It is easy to make the mistake of duplicating json struct field
tags especially when copy/pasting. This commit causes go vet to
report the mistake. Only field tags in the same struct type are
considered, because that is the only case which is undoubtedly an
error. (e.g. an embedded struct may have a tag name that conflicts
with a parent struct, but that is valid and sometimes useful
because of the behavior of the encoding/json package).

Note that changing encoding/json to return an error in the case of
duplicate tag names would violate backwards compatibility, and is
therefore not an option for go 1.

Fixes #12791

Change-Id: I4130e4c04b177694cc0daf8f1acaf0751d4f062b