Update patch set 1
Patch Set 1: Code-Review+2

I'm taking Mathieu's +1 as a +2, since this is blocking work we're doing today here in Paris on a Sunday. Don't want to wait for Sydney or the US to get to work.

The Go 1.5 build tag was added by Minux 3 months before the Go 1.5 release, and we're also now at 3 months before the Go 1.6 release, so it seems like a good time.

Patch-set: 1
Label: Code-Review=+2