Update patch set 3
Patch Set 3:

Alex, If my console is set to "Raster Fonts" it doesn't "fix" it. But if it is set to "Lucida Console" it does "fix" it.

However, it looks like the correct thing to do is to call WriteConsole if attached. I'll upload a revision with that change.

Patch-set: 3
Label: Code-Review=0