Update patch set 4
Patch Set 4:

Thanks very much for working on this. Let's find a way to do it without bundling the entire package utf16 and utf8 into the runtime. In particular, there is already a utf8 decoder in the runtime; we should not need a second. The utf16 code may be unavoidable, but even there I think it would be reasonable to use bundle once and delete the code you don't need. That code will literally never change.

The utf8 dependency can be removed by using runes := []rune(b). This would include any final partial rune at the end of b, but since write cannot return a partial result without being treated as an error, this is probably okay. In practice the runtime should not be sending partial utf8 bytes to begin with.

But more importantly, I do not think we want print allocating memory at all. For example, when malloc is broken it needs to print errors. It can't do that if print mallocs. So while you could avoid the utf8 dependency, there would still be two allocation+copy loops here: one for the utf8->rune conversion and one for the rune->utf16 conversion.

I think that instead we want a smallish buffer, say [1000]uint16, with a lock around it, and writeConsole would acquire the lock, use the buffer, release the lock. Locks during printing are a bit worrisome too, but it's at least better than allocating. You can loop over the input one rune at a time by calling charntorune, and when the 1000-uint16 buffer has fewer than 2 elements left, flush it. To convert from rune to utf16 you only need utf16.EncodeRune, which is two lines and not worth importing. The relevant constants are already defined in the runtime. 

I think the entire conversion might boil down to something without any allocations nor package bundling, like:

	const surr2 = (surrogateMin+surrogateMax+1)/2
	
	w := 0
	for len(b) > 0 {
		if w >= cap(utf16tmp) - 2 {
			writeConsoleUTF16(utf16tmp[:w])
			w = 0
		}
		r, n := charntorune(b)
		b = b[n:]
		if r < 0x10000 {
			utf16tmp = utf16tmp[:w+1]
			utf16tmp[w] = uint16(r)
		} else {
			utf16tmp = utf16tmp[:w+2]
			utf16tmp[w] = surrogateMin + uint16(r>>10)&0x3ff
			utf16tmp[w+1] = surr2 + uint16(r)&0x3ff
		}
	}
	writeConsoleUTF16(utf16tmp[:w])

Patch-set: 4