runtime: never allocate new M when jumping time forward
When we're jumping time forward, it means everyone is asleep, so there
should always be an M available. Furthermore, time jumping is only
used in nacl, where we can't allocate a new M anyway. More
furthermore, this causes both allocation and write barriers in
contexts that may be running without a P (such as in sysmon).

Hence, replace this allocation with a throw.

Updates #10600.

Change-Id: I2cee70d5db828d0044082878995949edb25dda5f