Update patch set 54
Patch Set 54:

(4 comments)

Matthew and Dave,

The relevant issue is #11647 (https://github.com/golang/go/issues/11647). I forgot to link to it before submitting my change.

There are two primary goals that I have making these changes. The first is (along with other refactorings) to improve the readability and accessibility of the runtime. The runtime should eventually (mostly) look like a Go program rather than a C program. The second is to improve the modularity of the runtime. That would help with stubbing out parts of the runtime when trying experiments. Josh mentions some more benefits on his comment on the issue.

It's true that we're introducing some complexity in this change, but I think it's pretty small for what we're getting. The bulk of this change is in moved files. Another large part of the line-count was in generated files such as cmd/compile/internal/gc/y.go and cmd/dist/deps.go. The non-trivial parts of the change included a small implementation change for arm atomics to remove the dependency on the runtime's mutex, a few couple of line changes to cmd/dist, cmd/go, cmd/compile, and updating some files such as deps_test.go to mention this new package.

Patch-set: 54