Update patch set 54
Patch Set 54:

> There are two primary goals that I have making these changes.

In addition to what Michael wrote, I want to add that most of the complexity here in the surrounding pieces (cmd/go, cmd/dist, cmd/compile) is a one-time cost. It's true that that cost is not worth paying if all we do is move atomic operations out of the runtime, but it certainly seems worth paying to get a runtime that is not one giant unstructured directory.

With respect to the freeze, the goal of this CL was to do a trivial bit of split-up now so that we have a whole freeze period and release to understand the implications, with the luxury that the change is small enough we can just roll it back if there is some major surprise.

Assuming Go 1.6 is released with this slightly split-up runtime, then we will have more confidence that splitting up the runtime further for Go 1.7 won't need to be rolled back.

Patch-set: 54