Update patch set 54
Patch Set 54:

> I'd like to see a design docs for #11647,
 > or at least what should the final runtime
 > structure look like. Thanks.

FWIW, we know we don't know what the final runtime structure will look like, but we also know that one big unstructured directory is not the answer. So while you're right that we need a design for some of the more difficult interconnections, we don't need a design to make sure that we can introduce some basic package splits.

Patch-set: 54